Fixed issue #1031.

+++ src/de/matthiasmann/twl/util/PNGSizeDeterminer.java
@@ -525,7 +525,8 @@ public class PNGSizeDeterminer {
                 throw new EOFException();
             }
             amount -= skipped;
-            size += skipped;
+            // this downcast should(tm) work
+            size += (int) skipped;
         }
     }
